renamed data directory to api
