+++ server/api/ArtistAPI.ts
